Split Widgets class to org.jline.widget package
